requirements.txt
# Core dependencies
python-dotenv==1.0.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# OCR and Document Processing
boto3==1.28.62  # AWS SDK for Textract
pytesseract==0.3.10
Pillow==10.4.0

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.14.0
opencv-python==4.8.1.78

# Facial Recognition
deepface==0.0.79

# Database and Storage
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# API Clients
requests==2.31.0
twilio==8.10.0
sendgrid==6.10.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Added from the code block
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4 